tool: support for `ReVanced`
* Custom tool sources
* YouTube Music included
* Unified config for ReVanced & ReVanced Extended

+++ action.sh
@@ -3,9 +3,9 @@
 TEMP_DIR="temp"
 BUILD_DIR="build"
 
-if [ "${TOKEN:-}" ]; then GH_HEADER="Authorization: token ${TOKEN}"; else GH_HEADER=; fi
+if [ "${GITHUB_TOKEN:-}" ]; then GH_HEADER="Authorization: token ${GITHUB_TOKEN}"; else GH_HEADER=; fi
 NEXT_VER_CODE=${NEXT_VER_CODE:-$(date +'%Y%m%d')}
-REBUILD=false
+REBUILD=${REBUILD:-false}
 OS=$(uname -o)
 
 # -------------------- json/toml --------------------
@@ -37,54 +37,43 @@ abort() {
 }
 
 get_prebuilts() {
-	pr "Getting prebuilts"
-	local rv_cli_url rv_integrations_url rv_patches rv_patches_changelog rv_patches_dl rv_patches_url rv_integrations_rel rv_patches_rel
-	rv_cli_url=$(gh_req "https://api.github.com/repos/inotia00/revanced-cli/releases/latest" - | json_get 'browser_download_url') || return 1
-	RV_CLI_JAR="${PREBUILTS_DIR}/${rv_cli_url##*/}"
-	log "**CLI**: _${rv_cli_url##*/}_"
-        rv_integrations_rel="https://api.github.com/repos/inotia00/revanced-integrations/releases/latest"
-	rv_patches_rel="https://api.github.com/repos/inotia00/revanced-patches/releases/latest"
+	local integrations_src=$1 patches_src=$2 cli_src=$3
+	local patches_dir=${patches_src%/*}
+	patches_dir=${TEMP_DIR}/${patches_dir//[^[:alnum:]]/}-rv
+	local integrations_dir=${integrations_src%/*}
+	integrations_dir=${TEMP_DIR}/${integrations_dir//[^[:alnum:]]/}-rv
+	local cli_dir=${cli_src%/*}
+	cli_dir=${TEMP_DIR}/${cli_dir//[^[:alnum:]]/}-rv
+	mkdir -p "$patches_dir" "$integrations_dir" "$cli_dir"
+
+	pr "Getting prebuilts (${patches_src%/*})" >&2
+	local rv_cli_url rv_integrations_url rv_patches rv_patches_dl rv_patches_url rv_patches_json
+
+	rv_cli_url=$(gh_req "https://api.github.com/repos/${cli_src}/releases/latest" - | json_get 'browser_download_url') || return 1
+	local rv_cli_jar="${cli_dir}/${rv_cli_url##*/}"
+	echo "\n**CLI**: $(cut -d/ -f4 <<<"$rv_cli_url")/$(cut -d/ -f9 <<<"$rv_cli_url")" >>"$patches_dir/changelog.md"
+
+	local rv_integrations_rel="https://api.github.com/repos/${integrations_src}/releases/latest"
+	local rv_patches_rel="https://api.github.com/repos/${patches_src}/releases/latest"
+	
 	rv_integrations_url=$(gh_req "$rv_integrations_rel" - | json_get 'browser_download_url')
-	RV_INTEGRATIONS_APK="${PREBUILTS_DIR}/${rv_integrations_url##*/}"
-	log "**Integrations**: _${rv_integrations_url##*/}_"
+	local rv_integrations_apk="${integrations_dir}/${rv_integrations_url##*/}"
+	echo "**Integrations**: $(cut -d/ -f4 <<<"$rv_integrations_url")/$(cut -d/ -f9 <<<"$rv_integrations_url")" >>"$patches_dir/changelog.md"
 
 	rv_patches=$(gh_req "$rv_patches_rel" -)
-	rv_patches_changelog=$(json_get 'body' <<<"$rv_patches" | sed 's/\(\\n\)\+/\\n/g')
 	rv_patches_dl=$(json_get 'browser_download_url' <<<"$rv_patches")
-	RV_PATCHES_JSON="${PREBUILTS_DIR}/patches-$(json_get 'tag_name' <<<"$rv_patches").json"
+	rv_patches_json="${patches_dir}/patches-$(json_get 'tag_name' <<<"$rv_patches").json"
 	rv_patches_url=$(grep 'jar' <<<"$rv_patches_dl")
-	RV_PATCHES_JAR="${PREBUILTS_DIR}/${rv_patches_url##*/}"
-	[ -f "$RV_PATCHES_JAR" ] || REBUILD=true
-	log "**Patches**: _${rv_patches_url##*/}_\n"
-        log "**Changelog**:\n[Revanced](https://github.com/revanced/revanced-patches/releases/) | [Revanced-Extended](https://github.com/inotia00/revanced-patches/releases/)\n"
+	local rv_patches_jar="${patches_dir}/${rv_patches_url##*/}"
+	[ -f "$rv_patches_jar" ] || REBUILD=true
+	echo "**Patches**: $(cut -d/ -f4 <<<"$rv_patches_url")/$(cut -d/ -f9 <<<"$rv_patches_url")" >>"$patches_dir/changelog.md"
 
-	dl_if_dne "$RV_CLI_JAR" "$rv_cli_url"
-	dl_if_dne "$RV_INTEGRATIONS_APK" "$rv_integrations_url"
-	dl_if_dne "$RV_PATCHES_JAR" "$rv_patches_url"
-	dl_if_dne "$RV_PATCHES_JSON" "$(grep 'json' <<<"$rv_patches_dl")"
+	dl_if_dne "$rv_cli_jar" "$rv_cli_url" >&2
+	dl_if_dne "$rv_integrations_apk" "$rv_integrations_url" >&2
+	dl_if_dne "$rv_patches_jar" "$rv_patches_url" >&2
+	dl_if_dne "$rv_patches_json" "$(grep 'json' <<<"$rv_patches_dl")" >&2
 
-	HTMLQ="${TEMP_DIR}/htmlq"
-	if [ ! -f "${TEMP_DIR}/htmlq" ]; then
-         	req "https://github.com/mgdm/htmlq/releases/latest/download/htmlq-x86_64-linux.tar.gz" "${TEMP_DIR}/htmlq.tar.gz"
-		tar -xf "${TEMP_DIR}/htmlq.tar.gz" -C "$TEMP_DIR"
-		rm "${TEMP_DIR}/htmlq.tar.gz"
-	fi
-}
-
-set_prebuilts() {
-	[ -d "$PREBUILTS_DIR" ] || abort "${PREBUILTS_DIR} directory could not be found"
-	RV_CLI_JAR=$(find "$PREBUILTS_DIR" -maxdepth 1 -name "revanced-cli-*.jar" | tail -n1)
-	[ "$RV_CLI_JAR" ] || abort "revanced cli not found"
-	log "**CLI**: _${RV_CLI_JAR#"$PREBUILTS_DIR/"}_"
-	RV_INTEGRATIONS_APK=$(find "$PREBUILTS_DIR" -maxdepth 1 -name "revanced-integrations-*.apk" | tail -n1)
-	[ "$RV_INTEGRATIONS_APK" ] || abort "revanced integrations not found"
-	log "**Integrations**: _${RV_INTEGRATIONS_APK#"$PREBUILTS_DIR/"}_"
-	RV_PATCHES_JAR=$(find "$PREBUILTS_DIR" -maxdepth 1 -name "revanced-patches-*.jar" | tail -n1)
-	[ "$RV_PATCHES_JAR" ] || abort "revanced patches not found"
-	log "**Patches**: _${RV_PATCHES_JAR#"$PREBUILTS_DIR/"}_"
-	RV_PATCHES_JSON=$(find "$PREBUILTS_DIR" -maxdepth 1 -name "patches-*.json" | tail -n1)
-	[ "$RV_PATCHES_JSON" ] || abort "patches.json not found"
-	HTMLQ="${TEMP_DIR}/htmlq"
+	echo "$rv_cli_jar" "$rv_integrations_apk" "$rv_patches_jar" "$rv_patches_json"
 }
 
 _req() {
@@ -93,14 +82,18 @@ _req() {
 	else
 		local dlp
 		dlp="$(dirname "$2")/tmp.$(basename "$2")"
-		wget -nv -O "$dlp" --header="$3" "$1"
+		if [ -f "$dlp" ]; then
+			while [ -f "$dlp" ]; do sleep 1; done
+			return
+		fi
+		wget -nv -O "$dlp" --header="$3" "$1" || return 1
 		mv -f "$dlp" "$2"
 	fi
 }
 req() { _req "$1" "$2" "User-Agent: Mozilla/5.0 (X11; Linux x86_64; rv:108.0) Gecko/20100101 Firefox/108.0"; }
 gh_req() { _req "$1" "$2" "$GH_HEADER"; }
 
-log() { echo -e "$1  " >>build.md; }
+log() { echo -e "$1  " >>"build.md"; }
 get_largest_ver() {
 	local vers m
 	vers=$(tee)
@@ -113,13 +106,25 @@ semver_validate() {
 	[ ${#ac} = 0 ]
 }
 get_patch_last_supported_ver() {
-        jq -r ".[]
-		| select(.compatiblePackages[].name==\"${1}\" and .excluded==false)
-		| .compatiblePackages[].versions" "$RV_PATCHES_JSON" |
-	tr -d ' ,\t[]"' | grep -v '^$' | sort | uniq -c | sort -nr | head -1 | xargs | cut -d' ' -f2 || return 1
+	local inc_sel exc_sel
+	inc_sel=$(list_args "$2" | sed 's/.*/\.name == "&"/' | paste -sd '~' | sed 's/~/ or /g' || :)
+	exc_sel=$(list_args "$3" | sed 's/.*/\.name != "&"/' | paste -sd '~' | sed 's/~/ and /g' || :)
+	inc_sel=${inc_sel:-false}
+	if [ "$4" = false ]; then inc_sel="${inc_sel} or .excluded==false"; fi
+	jq -r ".[]
+			| .name |= ascii_downcase | .name |= gsub(\"\\\\s\";\"-\")
+			| select(.compatiblePackages[].name==\"${1}\")
+			| select(${inc_sel})
+			| select(${exc_sel:-true})
+			| .compatiblePackages[].versions" "$5" |
+		tr -d ' ,\t[]"' | grep -v '^$' | sort | uniq -c | sort -nr | head -1 | xargs | cut -d' ' -f2 || return 1
 }
 
 dl_if_dne() {
+	[ "${DRYRUN:-}" ] && {
+		: >"$1"
+		return 0
+	}
 	if [ ! -f "$1" ]; then
 		pr "Getting '$1' from '$2'"
 		req "$2" "$1"
@@ -137,15 +142,15 @@ isoneof() {
 dl_apkmirror() {
 	local url=$1 version=${2// /-} output=$3 apkorbundle=$4 arch=$5 dpi=$6
 	[ "${DRYRUN:-}" ] && {
-		echo >"$output"
+		: >"$output"
 		return 0
 	}
 	local resp node app_table dlurl=""
 	[ "$arch" = universal ] && apparch=(universal noarch 'arm64-v8a + armeabi-v7a') || apparch=("$arch")
 	url="${url}/${url##*/}-${version//./-}-release/"
 	resp=$(req "$url" -) || return 1
-	for ((n = 2; n < 40; n++)); do
-		node=$($HTMLQ "div.table-row:nth-child($n)" -r "span:nth-child(n+3)" <<<"$resp")
+	for ((n = 1; n < 40; n++)); do
+		node=$($HTMLQ "div.table-row.headerFont:nth-last-child($n)" -r "span:nth-child(n+3)" <<<"$resp")
 		if [ -z "$node" ]; then break; fi
 		app_table=$($HTMLQ --text --ignore-whitespace <<<"$node")
 		if [ "$(sed -n 3p <<<"$app_table")" = "$apkorbundle" ] && { [ "$apkorbundle" = BUNDLE ] ||
@@ -183,9 +188,9 @@ get_apkmirror_pkg_name() { req "$1" - | sed -n 's;.*id=\(.*\)" class="accent_col
 # --------------------------------------------------
 
 patch_apk() {
-	local stock_input=$1 patched_apk=$2 patcher_args=$3
+	local stock_input=$1 patched_apk=$2 patcher_args=$3 rv_cli_jar=$4 rv_patches_jar=$5
 	declare -r tdir=$(mktemp -d -p $TEMP_DIR)
-	local cmd="java -jar $RV_CLI_JAR --rip-lib x86_64 --rip-lib x86 --temp-dir=$tdir -c -a $stock_input -o $patched_apk -b $RV_PATCHES_JAR --keystore=revanced.keystore $patcher_args"
+	local cmd="java -jar $rv_cli_jar --rip-lib x86_64 --rip-lib x86 --temp-dir=$tdir -c -a $stock_input -o $patched_apk -b $rv_patches_jar --keystore=revanced.keystore $patcher_args"
 	pr "$cmd"
 	if [ "${DRYRUN:-}" = true ]; then
 		cp -f "$stock_input" "$patched_apk"
@@ -196,13 +201,17 @@ patch_apk() {
 }
 
 build_rv() {
-	local -n args=$1
+	eval "declare -A args=${1#*=}"
 	local version build_mode_arr pkg_name
 	local mode_arg=${args[build_mode]} version_mode=${args[version]}
 	local app_name=${args[app_name]}
 	local app_name_l=${app_name,,}
+	app_name_l=${app_name_l// /-}
+	local table=${args[table]}
 	local dl_from=${args[dl_from]}
 	local arch=${args[arch]}
+	if [ "$arch" = 'universal' ]; then local arch_f="all"; else local arch_f="$arch"; fi
+
 	local p_patcher_args=()
 	p_patcher_args+=("$(join_args "${args[excluded_patches]}" -e) $(join_args "${args[included_patches]}" -i)")
 	[ "${args[exclusive_patches]}" = true ] && p_patcher_args+=("--exclusive")
@@ -215,79 +224,85 @@ build_rv() {
 	if [ "$version_mode" = auto ]; then
 		version=$(
 			get_patch_last_supported_ver "$pkg_name" \
-				"${args[included_patches]}" "${args[excluded_patches]}" "${args[exclusive_patches]}"
+				"${args[included_patches]}" "${args[excluded_patches]}" "${args[exclusive_patches]}" "${args[ptjs]}"
 		) || get_latest_ver=true
+	elif isoneof "$version_mode" latest beta; then
+		get_latest_ver=true
+		p_patcher_args+=("--experimental")
+	else
+		version=$version_mode
+		p_patcher_args+=("--experimental")
 	fi
 	if [ $get_latest_ver = true ]; then
 		local apkmvers
 		if [ "$dl_from" = apkmirror ]; then
-			apkmvers=$(get_apkmirror_vers "${args[apkmirror_dlurl]##*/}" "false")
-			version=$(get_largest_ver <<<"$apkmvers") || version=$(head -1 <<<"$apkmvers")
+		apkmvers=$(get_apkmirror_vers "${args[apkmirror_dlurl]##*/}" "false")
+		version=$(get_largest_ver <<<"$apkmvers") || version=$(head -1 <<<"$apkmvers")
 		fi
 	fi
 	if [ -z "$version" ]; then
-		epr "empty version, not building ${app_name}."
+		epr "empty version, not building ${table}."
 		return 0
 	fi
-	pr "Choosing version '${version}' (${app_name})"
+	pr "Choosing version '${version}' for ${table}"
 	local version_f=${version// /}
-	local stock_apk="${TEMP_DIR}/${pkg_name}-${version_f}-${arch}.apk"
+	version_f=${version_f#v}
+	local stock_apk="${TEMP_DIR}/${pkg_name}-${version_f}-${arch_f}.apk"
 	if [ ! -f "$stock_apk" ]; then
-		if [ "$dl_from" = apkmirror ]; then
-			pr "Downloading '${app_name}' from APKMirror"
-			local apkm_arch
-			if [ "$arch" = "all" ]; then
-				apkm_arch="universal"
-			elif [ "$arch" = "arm64-v8a" ]; then
-				apkm_arch="arm64-v8a"
-			elif [ "$arch" = "arm-v7a" ]; then
-				apkm_arch="armeabi-v7a"
-			fi
-			if ! dl_apkmirror "${args[apkmirror_dlurl]}" "$version" "$stock_apk" APK "$apkm_arch" "${args[dpi]}"; then
-				epr "ERROR: Could not find any release of '${app_name}' with version '${version}', arch '${apkm_arch}' and dpi '${args[dpi]}' from APKMirror"
-				return 0
-			fi
+		if [ -z "${args[apkmirror_dlurl]}" ]; then continue; fi
+		pr "Downloading '${table}' from APKMirror"
+		local apkm_arch
+		if [ "$arch" = "universal" ]; then
+			apkm_arch="universal"
+		elif [ "$arch" = "arm64-v8a" ]; then
+			apkm_arch="arm64-v8a"
+		elif [ "$arch" = "arm-v7a" ]; then
+			apkm_arch="armeabi-v7a"
 		fi
+		if ! dl_apkmirror "${args[apkmirror_dlurl]}" "$version" "$stock_apk" APK "$apkm_arch" "${args[dpi]}"; then
+			epr "ERROR: Could not find any release of '${table}' with version '${version}', arch '${apkm_arch}' and dpi '${args[dpi]}' from APKMirror"
+			continue
+		fi
+		if [ ! -f "$stock_apk" ]; then return 0; fi
 	fi
-	grep -q "${app_name}:" build.md || log "${app_name}: ${version}"
-	if [ "${args[merge_integrations]}" = true ]; then
-		p_patcher_args+=("-m ${RV_INTEGRATIONS_APK}")
-	fi
+	log "${table}: ${version}"
 
+	if [ "${args[merge_integrations]}" = true ]; then p_patcher_args+=("-m ${args[integ]}"); fi
 	local microg_patch
-	microg_patch=$(jq -r ".[] | select(.compatiblePackages[].name==\"${pkg_name}\") | .name" "$RV_PATCHES_JSON" | grep -F microg || :)
+	microg_patch=$(jq -r ".[] | select(.compatiblePackages[].name==\"${pkg_name}\") | .name" "${args[ptjs]}" | grep -iF microg || :)
 	if [ "$microg_patch" ]; then
+		microg_patch="${microg_patch,,}"
+		microg_patch="${microg_patch// /-}"
 		p_patcher_args=("${p_patcher_args[@]//-[ei] ${microg_patch}/}")
 	fi
 
-	local stock_bundle_apk="${TEMP_DIR}/${pkg_name}-${version_f}-${arch}-bundle.apk"
+	local stock_bundle_apk="${TEMP_DIR}/${pkg_name}-${version_f}-${arch_f}-bundle.apk"
 	local is_bundle=false
-	if [ "$mode_arg" = apk ]; then
-		build_mode_arr=(apk)
-	fi
+	build_mode_arr=(apk)
 	local patcher_args patched_apk build_mode
+	local rv_brand_f=${args[rv_brand],,}
+	rv_brand_f=${rv_brand_f// /-}
 	for build_mode in "${build_mode_arr[@]}"; do
 		patcher_args=("${p_patcher_args[@]}")
-		pr "Building '${app_name}' (${arch}) in '$build_mode' mode"
+		pr "Building '${table}' in '$build_mode' mode"
 		if [ "$microg_patch" ]; then
-			patched_apk="${TEMP_DIR}/${app_name_l}-${RV_BRAND_F}-${version_f}-${arch}-${build_mode}.apk"
+			patched_apk="${TEMP_DIR}/${app_name_l}-${rv_brand_f}-${version_f}-${arch_f}-${build_mode}.apk"
 			patcher_args+=("-i ${microg_patch}")
 		else
-			patched_apk="${TEMP_DIR}/${app_name_l}-${RV_BRAND_F}-${version_f}-${arch}.apk"
+			patched_apk="${TEMP_DIR}/${app_name_l}-${rv_brand_f}-${version_f}-${arch_f}.apk"
 		fi
 		if [ ! -f "$patched_apk" ] || [ "$REBUILD" = true ]; then
-			if ! patch_apk "$stock_apk" "$patched_apk" "${patcher_args[*]}"; then
-				pr "Building '${app_name}' failed!"
+			if ! patch_apk "$stock_apk" "$patched_apk" "${patcher_args[*]}" "${args[cli]}" "${args[ptjar]}"; then
+				epr "Building '${table}' failed!"
 				return 0
 			fi
 		fi
 		if [ "$build_mode" = apk ]; then
-			local apk_output="${BUILD_DIR}/${app_name_l}-${RV_BRAND_F}-v${version_f}-${arch}.apk"
+			local apk_output="${BUILD_DIR}/${app_name_l}-${rv_brand_f}-v${version_f}-${arch_f}.apk"
 			cp -f "$patched_apk" "$apk_output"
-			pr "Built ${app_name} (${arch}) (non-root): '${apk_output}'"
+			pr "Built ${table} (non-root): '${apk_output}'"
 			continue
 		fi
-		pr "Built ${app_name} (${arch})"
 	done
 }
